requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
rich>=13.7.0
fastapi>=0.111.0
uvicorn>=0.29.0
httpx>=0.27.0
python-multipart>=0.0.9
itsdangerous>=2.1.2
jinja2>=3.1.4
